requirements_docker_gpu.txt
# local package
-e .

# external requirements
click==8.0.3
Sphinx
coverage
awscli
flake8
python-dotenv==0.19.2
matplotlib==3.4.3
numpy==1.21.2
setuptools==60.2.0
# torch==1.10.1
tqdm==4.62.3
hydra-core==1.1.1
omegaconf==2.1.1